and moved to pkg/resources
