+++ pkg/reconciler/v1alpha1/service/resources/route.go
@@ -24,8 +24,8 @@ import (
 	"github.com/knative/pkg/kmeta"
 	"github.com/knative/serving/pkg/apis/serving"
 	"github.com/knative/serving/pkg/apis/serving/v1alpha1"
-	"github.com/knative/serving/pkg/reconciler/resources"
 	"github.com/knative/serving/pkg/reconciler/v1alpha1/service/resources/names"
+	"github.com/knative/serving/pkg/resources"
 )
 
 // MakeRoute creates a Route from a Service object.
